+++ src/entities/userInfo/model/userInfoStore.tsx
@@ -1,5 +1,6 @@
 import { create } from 'zustand'
 import { UserInfo } from '../api/types'
+import { persist } from 'zustand/middleware'
 
 interface UserInfoStore {
   userInfo: UserInfo | null
@@ -7,8 +8,15 @@ interface UserInfoStore {
   resetUserInfo: () => void
 }
 
-export const useUserInfoStore = create<UserInfoStore>()((set) => ({
-  userInfo: null,
-  setUserInfo: (userInfo: UserInfo | null) => set({ userInfo }),
-  resetUserInfo: () => set({ userInfo: null }),
-}))
+export const useUserInfoStore = create<UserInfoStore>()(
+  persist(
+    (set) => ({
+      userInfo: null,
+      setUserInfo: (userInfo: UserInfo | null) => set({ userInfo }),
+      resetUserInfo: () => set({ userInfo: null }),
+    }),
+    {
+      name: 'user-storage',
+    },
+  ),
+)
